requirements.txt
environs==11.0.0
protobuf
keelson==0.3.1-pre.20
pyserial
eclipse-zenoh==0.10.1-rc0
opencv-python==4.10.0.84